Reorganiza exemplos de templates do mogi
